pyproject.toml
[project]
name = "feature"
version = "0.1.0"
description = "RAG with Knowledge Graph and Concept Graph"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "nebula3-python>=3.5.0",
    "pymilvus>=2.3.0",
    "numpy>=1.24.0",
    "spacy>=3.7.0",
    "networkx>=3.2.0",
    "python-louvain>=0.16",
    "leidenalg>=0.9.0; platform_system != 'Windows'",
    "igraph>=0.10.0; platform_system != 'Windows'",
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "tiktoken>=0.5.0",
    "aiofiles>=23.2.0",
    " Pillow>=10.0.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.21",
    "pandas>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
